Fixes and code refactoring
## 1. Fix getProjectOrder and getContextOrder
Fix the case, when there were not context or project fields in state.
Now it returns empty list.

## 2. Rearrange some code in order.js
Make addId() function simpler

## 3. Change names for projectOrder and contextOrder selectors
Change names to getProjectOrder and getContextOrder to match the common
pattern for selectors

##4. Refactor getOrderedProjectsList and getOrderedContextsList
  + Change variables names to make them more readable (mapForOrdering
to projects and contexts accordingly).
  + Also fix a problem with empty order array. Initially it returned
the empty array. Now it transforms project or context maps into lists.