Fix local storage saivings
Now it save to local storage only if user wasn’t logged in.
Also, when user logs out it tries to load state from local storage.
Otherwise it set state to empty initial state
Finally, when user tries to log in, and there is no data fetched from
firebase, it assumed that it is new account and tries to populate state
from local storage. If not, it just sets it to empty initial state